Sprint-4
In this sprint 4 we have done the connection to the nodes to the Watson IOT and debug and showing it to the dashboard. Also, in this we are creating an app and then
collaborating it to the node red and web UI. Then running it in the mobile and accessing it.